11. Capstone Project/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi
uvicorn
torch
torchvision
pillow
requests
python-multipart